+++ lib/functions-rewrites.php
@@ -2,6 +2,20 @@
 /** TODO: Refactor this to use an array rather than repeating code.
  */
 
+add_action( 'init', 'red_flag_rewrites' );
+/**
+ * Redirects /red-flags to /category/articles/red-flags/
+ */
+function red_flag_rewrites() {
+  $cat = get_category_by_path( 'articles/red-flags' );
+  if ( $cat ) {
+    add_rewrite_rule(
+        '^red-flags/?$',
+        'index.php?category_name=articles/' . $cat->slug,
+        'top'
+    );
+  }
+}
 add_action( 'init', 'committed_rewrites' );
 /**
  * Redirects /committed to /category/audio/committed.