+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- Redirect to red-flags catagory https://novaramedia.com/redflags to https://novaramedia.com/category/articles/red-flags/
+
 ## [4.2.8] - 2025-04-30
 
 ### Fixed