+++ CHANGELOG.md
@@ -8,11 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+
 - Redirect from https://novaramedia.com/asksophie to this google form https://docs.google.com/forms/d/e/1FAIpQLSegJV5jED2FhIUS7_rryZC6V2Y65W7W-kE3tXGMH7zr4sl_uQ/viewform
 
 ## [4.2.9] - 2025-05-21
 
 ### Added
+
 - Redirect to red-flags category https://novaramedia.com/redflags to https://novaramedia.com/category/articles/red-flags/
 
 ## [4.2.8] - 2025-04-30